slightly refactored
