+++ packages/typescript-typings/types/trezor-connect/index.d.ts
@@ -254,8 +254,8 @@ declare module 'trezor-connect' {
 
     export interface Address {
         address: string;
-        path: number[];
-        serializedPath: string;
+        path: number[]; // TODO: is string in accounts.addresses
+        serializedPath: string; // TODO: doesnt exist in accounts.addresses
     }
 
     export interface ComposeTransactionParams extends CommonParams {
